# perf: bundle `@rspack/core` for faster startup