--- a/packages/rspack-test-tools/tests/hookCases/compilation#chunkHash/update-hash/hooks.snap.txt
+++ b/packages/rspack-test-tools/tests/hookCases/compilation#chunkHash/update-hash/hooks.snap.txt
@@ -4,7 +4,7 @@
 
 ```javascript
 Array [
-  Chunk {
+  _Chunk {
     "auxiliaryFiles": Set {},
     "chunkReason": undefined,
     "contentHash": Object {},