--- a/packages/rspack/src/builtin-plugin/css-extract/loader.ts
+++ b/packages/rspack/src/builtin-plugin/css-extract/loader.ts
@@ -1,8 +1,7 @@
 import path from "node:path";
 
 import type { Filename, LoaderContext, LoaderDefinition } from "../..";
-import { CssExtractRspackPlugin } from "./index";
-import { stringifyLocal, stringifyRequest } from "./utils";
+import { PLUGIN_NAME, stringifyLocal, stringifyRequest } from "./utils";
 
 export const BASE_URI = "webpack://";
 export const MODULE_TYPE = "css/mini-extract";
@@ -251,7 +250,7 @@ export const pitch: LoaderDefinition["pitch"] = function (request, _, data) {
 			return "";
 		})();
 
-		let resultSource = `// extracted by ${CssExtractRspackPlugin.pluginName}`;
+		let resultSource = `// extracted by ${PLUGIN_NAME}`;
 
 		// only attempt hotreloading if the css is actually used for something other than hash values
 		resultSource +=
@@ -260,8 +259,7 @@ export const pitch: LoaderDefinition["pitch"] = function (request, _, data) {
 				: result;
 
 		if (dependencies.length > 0) {
-			parseMeta[CssExtractRspackPlugin.pluginName] =
-				JSON.stringify(dependencies);
+			parseMeta[PLUGIN_NAME] = JSON.stringify(dependencies);
 		}
 
 		callback(null, resultSource, undefined, data);