--- a/packages/rspack/src/builtin-plugin/css-extract/utils.ts
+++ b/packages/rspack/src/builtin-plugin/css-extract/utils.ts
@@ -6,6 +6,8 @@ export function isAbsolutePath(str: string) {
 	return path.posix.isAbsolute(str) || path.win32.isAbsolute(str);
 }
 
+export const PLUGIN_NAME = "css-extract-rspack-plugin";
+
 const RELATIVE_PATH_REGEXP = /^\.\.?[/\\]/;
 
 export function isRelativePath(str: string) {