--- a/packages/rspack/tsup.config.ts
+++ b/packages/rspack/tsup.config.ts
@@ -1,0 +1,7 @@
+import { defineConfig } from "tsup";
+
+export default defineConfig({
+	entry: ["./src/index.ts"],
+	format: ["cjs"],
+	target: "node16"
+});