--- a/packages/rspack/tsup.config.ts
+++ b/packages/rspack/tsup.config.ts
@@ -1,7 +1,24 @@
 import { defineConfig } from "tsup";
+import prebundleConfig from "./prebundle.config.mjs";
+
+const aliasCompiledPlugin = {
+	name: "alias-compiled-plugin",
+	setup(build) {
+		const { dependencies } = prebundleConfig;
+
+		for (const item of dependencies) {
+			const depName = typeof item === "string" ? item : item.name;
+			build.onResolve({ filter: new RegExp(`^${depName}$`) }, () => ({
+				path: `../compiled/${depName}/index.js`,
+				external: true
+			}));
+		}
+	}
+};
 
 export default defineConfig({
 	entry: ["./src/index.ts"],
 	format: ["cjs"],
-	target: "node16"
+	target: "node16",
+	esbuildPlugins: [aliasCompiledPlugin]
 });