<!--
  Thank you for submitting a pull request!

  We appreciate the time and effort you have invested in making these changes. Please ensure that you provide enough information to allow others to review your pull request.

  Upon submission, your pull request will be automatically assigned with reviewers.

  If you want to learn more about contributing to this project, please visit: https://github.com/web-infra-dev/rspack/blob/main/CONTRIBUTING.md.
-->

## Summary

- This PR uses `tsup` to bundle `@rspack/core`, this reduces the number of modules required by Node.js and speeds up Rspack startup time.
- Move runtime code to `src/runtime` folder for easier management.
- Remove some packages from `prebundle.config.js` because they can be bundled to `dist/index.js`.
- Once Rslib is ready, we will switch from tsup to Rslib.

> TODO: `tests/plugin-test/css-extract/stringifyLocal.test.js` and `packages/rspack/scripts/check-documentation-coverage.mjs` depend on internal dist files and should be fixed in separate PRs.

## Comparasion

`require('@rspack/core')` **become 30ms faster**:

- before (65~75ms):

```bash
hyperfine --warmup 3 'node ./node_modules/@rsbuild/core/bin/rsbuild.js build'
Benchmark 1: node ./node_modules/@rsbuild/core/bin/rsbuild.js build
  Time (mean ± σ):     345.4 ms ±  23.0 ms    [User: 385.2 ms, System: 65.3 ms]
  Range (min … max):   328.7 ms … 405.7 ms    10 runs
```

<img width="688" alt="Screenshot 2024-10-09 at 17 50 37" src="https://github.com/user-attachments/assets/6387cd1e-eb37-45c4-8853-48f3df2c5580">

- after (35~40ms):

```bash
hyperfine --warmup 3 'node ./node_modules/@rsbuild/core/bin/rsbuild.js build'
Benchmark 1: node ./node_modules/@rsbuild/core/bin/rsbuild.js build
  Time (mean ± σ):     308.8 ms ±   4.0 ms    [User: 350.4 ms, System: 58.6 ms]
  Range (min … max):   304.6 ms … 319.0 ms    10 runs
```

<img width="624" alt="Screenshot 2024-10-09 at 18 09 26" src="https://github.com/user-attachments/assets/eb09dcb0-2bd3-4155-9568-fe3e36b478fd">

## Checklist

<!--- Check and mark with an "x" -->

- [ ] Tests updated (or not required).
- [ ] Documentation updated (or not required).