@chenjiahan 2024-10-12
The devserver CI failed because we need to update the module path snapshots. I will update them after releasing a new Rspack version.

![image](https://github.com/user-attachments/assets/184d21ee-3004-49e1-8388-5d61456d9704)