@fi3ework 2024-10-12
> Remove some packages from prebundle.config.js because they can be bundled to dist/index.js.

Do we have guidelines for deciding when to use `prebundle` or direct bundle? I assume we should use `prebundle` when direct bundling isn't possible, as it can handle more edge cases.