@chenjiahan 2024-10-12
Currently we have two cases to use `prebundle`:

1. Use `prebundle` for chunk splitting. For example, we want to load `browserslist` on demand, so we can use prebundle to output a separate `browserslist/index.js` chunk. This can be removed after Rslib supports chunk splitting.
2. Use `prebundle` to bundle the types of thrid-party packages. For example, we need the types of zod and `prebundle` can handle it.